requirements.txt
django>=1.4.2
django-floppyforms>=1.0
django-tastypie>=0.9